[Refactor] 공통 consts Core 모듈화
