+++ .gitsecret/paths/mapping.cfg
@@ -1,1 +1,1 @@
-keewe-domain/src/main/resources/application-domain.yml:34fdeb0b03ccaa825f2942ab6ae78bbdc4df2c72f518f28fe622f8b9b46078d3
+keewe-domain/src/main/resources/application-domain.yml:fcef66e2566094c50f27458b9c9b6cc9bd7ffbfe53b3a220f9c90e76ce4b804e
