+++ keewe-api/src/main/java/ccc/keeweapi/config/security/jwt/JwtAuthenticationEntryPoint.java
@@ -27,7 +27,7 @@ public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
     private void commence(HttpServletResponse response, KeeweAuthException kex) throws IOException {
         response.setContentType("application/json;charset=UTF-8");
         response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-        log.info("kex {}", kex.getKeeweRtnConsts().name());
+        log.info("Auth Exception {}", kex.getKeeweRtnConsts().name());
 
         JSONObject body = new JSONObject();
         body.put("timestamp", LocalDateTime.now().toString());
