+++ keewe-core/src/test/java/ccc/keewecore/KeeweCoreApplicationTests.java
@@ -1,0 +1,13 @@
+package ccc.keewecore;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class KeeweCoreApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
