+++ keewe-domain/src/main/resources/schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS `user`
     updated_at          DATETIME(6)     NOT NULL,
 
     PRIMARY KEY(user_id)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
+    ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE IF NOT EXISTS `profile_photo`
 (
@@ -20,16 +20,17 @@ CREATE TABLE IF NOT EXISTS `profile_photo`
     updated_at          DATETIME(6)     NOT NULL,
 
     PRIMARY KEY (profile_photo_id)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
+    ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE IF NOT EXISTS `profile`
 (
     profile_id          BIGINT(20)      NOT NULL        AUTO_INCREMENT,
     user_id             BIGINT(20)      NOT NULL,
     nickname            VARCHAR(30)     NOT NULL,
-    link                VARCHAR(30)     NOT NULL        UNIQUE,
+    link                VARCHAR(30)     NOT NULL,
     privacy             VARCHAR(20)     NOT NULL,
     profile_photo_id    BIGINT(20),
+    profile_status      VARCHAR(30)     NOT NULL,
     deleted             BIT             NOT NULL,
     created_at          DATETIME(6)     NOT NULL,
     updated_at          DATETIME(6)     NOT NULL,
@@ -37,7 +38,7 @@ CREATE TABLE IF NOT EXISTS `profile`
     PRIMARY KEY (profile_id),
     FOREIGN KEY (user_id) REFERENCES `user`(user_id),
     FOREIGN KEY (profile_photo_id) REFERENCES  `profile_photo`(profile_photo_id)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
+    ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE IF NOT EXISTS `buddy`
 (
@@ -50,7 +51,7 @@ CREATE TABLE IF NOT EXISTS `buddy`
     PRIMARY KEY (buddy_id),
     FOREIGN KEY (followee_id) REFERENCES `profile`(profile_id),
     FOREIGN KEY (followee_id) REFERENCES `profile`(profile_id)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
+    ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE IF NOT EXISTS `favorite_activities`
 (
@@ -61,7 +62,7 @@ CREATE TABLE IF NOT EXISTS `favorite_activities`
 
     FOREIGN KEY (profile_id) REFERENCES `profile`(profile_id),
     CONSTRAINT `favorite_activities_constraint` UNIQUE (profile_id, activity)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
+    ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE IF NOT EXISTS `profile_link`
 (
@@ -75,4 +76,10 @@ CREATE TABLE IF NOT EXISTS `profile_link`
 
     PRIMARY KEY (profile_link_id),
     FOREIGN KEY (profile_id) REFERENCES `profile`(profile_id)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
+    ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
+
+INSERT INTO `user` (user_id, email, password, phone_number, status, deleted, created_at, updated_at)
+VALUES(0, 'test@keewe.com', 'pass', '01012341234', 'ONBOARDING', false, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP);
+
+INSERT INTO `profile` (profile_id, user_id, nickname, link, privacy, profile_photo_id, profile_status, deleted, created_at, updated_at)
+VALUES (0, 0,'', '', 'PUBLIC', null, 'LINK_NEEDED', false, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP);
