+++ settings.gradle
@@ -5,3 +5,4 @@ include 'keewe-infra'
 include 'keewe-push'
 include 'keewe-batch'
 include 'keewe-statistics'
+include 'keewe-core'
